add tests using vcr
