+++ .github/workflows/unittest.yml
@@ -13,8 +13,23 @@ jobs:
       matrix:
         python-version: [3.11]
     if: ${{ !contains(github.event.head_commit.message, '[ci skip]') }}
+    services:
+      db:
+        image: mysql:8
+        env:
+          MYSQL_ROOT_PASSWORD: root
+          MYSQL_DATABASE: archivesspace
+          MYSQL_USER: as
+          MYSQL_PASSWORD: as123
+        ports:
+          - 3307:3306
+        options: --health-cmd="mysqladmin ping" --health-interval=10s --health-timeout=5s --health-retries=3
     steps:
       - uses: actions/checkout@v4
+      - run: sudo apt-get install -y mysql-client
+      - name: Import SQL dump file
+        run: |
+          gzip -dc ./test/fixtures/test_data.sql.gz | mysql --host 127.0.0.1 --port 3307 -uroot -proot archivesspace
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v5
         with:
@@ -23,6 +38,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install -r requirements.txt
+          pip install -r test-requirements.txt
       - name: Test with unittest
         run: |
-          as_api="http://localhost:4567" as_un="admin" as_pw="admin" python -m unittest discover test
+          as_api="http://localhost:4567" as_un="admin" as_pw="admin" db_un="as" db_pw="as123" db_host="127.0.0.1" db_name="archivesspace" db_port="3307" python -m unittest
