+++ test-requirements.txt
@@ -1,0 +1,1 @@
+vcrpy~=7.0.0
