+++ test/fixtures/mergesubjects_testdata.csv
@@ -1,0 +1,2 @@
+aspace_subject_id,title,scope_note,Merge into,aspace_subject_id2
+2,title 2,,title 3,3
