+++ test/fixtures/vcr_cassettes/test_utilities/test_bad_connection.yaml
@@ -1,0 +1,40 @@
+interactions:
+- request:
+    body: password=[REDACTED]&expiring=False
+    headers:
+      Accept:
+      - application/json
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      Content-Length:
+      - '36'
+      Content-Type:
+      - application/x-www-form-urlencoded
+      User-Agent:
+      - ArchivesSnake/0.1
+    method: POST
+    uri: http://localhost:4567/users/bad_user/login
+  response:
+    body:
+      string: '{"error":"Login failed"}
+
+        '
+    headers:
+      Cache-Control:
+      - private, must-revalidate, max-age=0
+      Content-Length:
+      - '25'
+      Content-Type:
+      - application/json
+      Date:
+      - Sat, 28 Feb 2026 00:36:24 GMT
+      Server:
+      - Jetty(9.4.44.v20210927)
+      X-Content-Type-Options:
+      - nosniff
+    status:
+      code: 403
+      message: Forbidden
+version: 1
