+++ test/fixtures/vcr_cassettes/test_utilities/test_get_digobj.yaml
@@ -1,0 +1,38 @@
+interactions:
+- request:
+    body: null
+    headers:
+      Accept:
+      - application/json
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      User-Agent:
+      - ArchivesSnake/0.1
+      X-ArchivesSpace-Session:
+      - '[REDACTED]'
+    method: GET
+    uri: http://localhost:4567/repositories/2/digital_objects/3
+  response:
+    body:
+      string: '{"lock_version":1,"digital_object_id":"DOID_352","title":"Test DO 1","publish":true,"restrictions":false,"created_by":"admin","last_modified_by":"admin","create_time":"2026-02-27T23:18:07Z","system_mtime":"2026-02-27T23:20:16Z","user_mtime":"2026-02-27T23:20:15Z","suppressed":false,"is_slug_auto":true,"digital_object_type":"still_image","jsonmodel_type":"digital_object","external_ids":[],"subjects":[],"linked_events":[],"extents":[],"lang_materials":[],"dates":[],"external_documents":[],"rights_statements":[],"linked_agents":[],"file_versions":[{"lock_version":0,"file_uri":"https://example.com/1","publish":true,"created_by":"admin","last_modified_by":"admin","create_time":"2026-02-27T23:20:15Z","system_mtime":"2026-02-27T23:20:15Z","user_mtime":"2026-02-27T23:20:15Z","is_representative":true,"use_statement":"image-service","jsonmodel_type":"file_version","identifier":"9","link_uri":"https://example.com/1/2"},{"lock_version":0,"file_uri":"https://example.com/1/2","publish":true,"created_by":"admin","last_modified_by":"admin","create_time":"2026-02-27T23:20:15Z","system_mtime":"2026-02-27T23:20:15Z","user_mtime":"2026-02-27T23:20:15Z","use_statement":"video-service","jsonmodel_type":"file_version","is_representative":false,"identifier":"10"}],"classifications":[],"notes":[],"collection":[],"linked_instances":[],"metadata_rights_declarations":[],"uri":"/repositories/2/digital_objects/3","repository":{"ref":"/repositories/2"},"tree":{"ref":"/repositories/2/digital_objects/3/tree"},"representative_file_version":{"lock_version":0,"file_uri":"https://example.com/1","publish":true,"created_by":"admin","last_modified_by":"admin","create_time":"2026-02-27T23:20:15Z","system_mtime":"2026-02-27T23:20:15Z","user_mtime":"2026-02-27T23:20:15Z","is_representative":true,"use_statement":"image-service","jsonmodel_type":"file_version","identifier":"9","link_uri":"https://example.com/1/2"}}
+
+        '
+    headers:
+      Cache-Control:
+      - private, must-revalidate, max-age=0
+      Content-Length:
+      - '1904'
+      Content-Type:
+      - application/json
+      Date:
+      - Fri, 27 Feb 2026 23:27:04 GMT
+      Server:
+      - Jetty(9.4.44.v20210927)
+      X-Content-Type-Options:
+      - nosniff
+    status:
+      code: 200
+      message: OK
+version: 1
