+++ test/fixtures/vcr_cassettes/test_utilities/test_repo_search.yaml
@@ -1,0 +1,45 @@
+interactions:
+- request:
+    body: null
+    headers:
+      Accept:
+      - application/json
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      User-Agent:
+      - ArchivesSnake/0.1
+      X-ArchivesSpace-Session:
+      - '[REDACTED]'
+    method: GET
+    uri: http://localhost:4567/repositories/2/search?aq=%7B%22jsonmodel_type%22:%20%22advanced_query%22,%20%22query%22:%20%7B%22jsonmodel_type%22:%20%22boolean_query%22,%20%22op%22:%20%22AND%22,%20%22subqueries%22:%20%5B%7B%22jsonmodel_type%22:%20%22field_query%22,%20%22field%22:%20%22title%22,%20%22value%22:%20%22Test%20resource%200%22%7D%5D%7D%7D&type%5B%5D=resource&page=1
+  response:
+    body:
+      string: '{"page_size":10,"first_page":1,"last_page":1,"this_page":1,"offset_first":1,"offset_last":1,"total_hits":1,"results":[{"id":"/repositories/2/resources/5","uri":"/repositories/2/resources/5","title":"Test
+        resource 0","title_ws":"Test resource 0","primary_type":"resource","types":["resource"],"suppressed":false,"publish":false,"system_generated":false,"repository":"/repositories/2","level_enum_s":["collection"],"finding_aid_language_enum_s":["eng"],"finding_aid_script_enum_s":["Latn"],"portion_enum_s":["whole"],"extent_type_enum_s":["cubic_feet"],"language_enum_s":["eng"],"date_type_enum_s":["inclusive"],"label_enum_s":["creation"],"created_by":"admin","last_modified_by":"admin","user_mtime":"2026-02-27T23:18:05Z","system_mtime":"2026-02-27T23:18:07Z","create_time":"2026-02-27T23:18:05Z","dates":["1906-10-29
+        - 1907-03-03"],"level":"collection","extents":["1 --- cubic_feet"],"identifier":"EAD-0","identifier_ws":"EAD-0","restrictions":"false","ead_id":"EAD.0","is_slug_auto":true,"four_part_id":"EAD
+        0","langcode":["eng"],"json":"{\"lock_version\":0,\"title\":\"Test resource
+        0\",\"publish\":false,\"restrictions\":false,\"ead_id\":\"EAD.0\",\"created_by\":\"admin\",\"last_modified_by\":\"admin\",\"create_time\":\"2026-02-27T23:18:05Z\",\"system_mtime\":\"2026-02-27T23:18:07Z\",\"user_mtime\":\"2026-02-27T23:18:05Z\",\"suppressed\":false,\"is_slug_auto\":true,\"is_finding_aid_status_published\":true,\"id_0\":\"EAD\",\"id_1\":\"0\",\"level\":\"collection\",\"finding_aid_language\":\"eng\",\"finding_aid_script\":\"Latn\",\"jsonmodel_type\":\"resource\",\"external_ids\":[],\"subjects\":[],\"linked_events\":[],\"extents\":[{\"lock_version\":0,\"number\":\"1\",\"created_by\":\"admin\",\"last_modified_by\":\"admin\",\"create_time\":\"2026-02-27T23:18:05Z\",\"system_mtime\":\"2026-02-27T23:18:05Z\",\"user_mtime\":\"2026-02-27T23:18:05Z\",\"portion\":\"whole\",\"extent_type\":\"cubic_feet\",\"jsonmodel_type\":\"extent\"}],\"lang_materials\":[{\"lock_version\":0,\"created_by\":\"admin\",\"last_modified_by\":\"admin\",\"create_time\":\"2026-02-27T23:18:05Z\",\"system_mtime\":\"2026-02-27T23:18:05Z\",\"user_mtime\":\"2026-02-27T23:18:05Z\",\"jsonmodel_type\":\"lang_material\",\"notes\":[],\"language_and_script\":{\"lock_version\":0,\"created_by\":\"admin\",\"last_modified_by\":\"admin\",\"create_time\":\"2026-02-27T23:18:05Z\",\"system_mtime\":\"2026-02-27T23:18:05Z\",\"user_mtime\":\"2026-02-27T23:18:05Z\",\"language\":\"eng\",\"jsonmodel_type\":\"language_and_script\"}}],\"dates\":[{\"lock_version\":0,\"begin\":\"1906-10-29\",\"end\":\"1907-03-03\",\"created_by\":\"admin\",\"last_modified_by\":\"admin\",\"create_time\":\"2026-02-27T23:18:05Z\",\"system_mtime\":\"2026-02-27T23:18:05Z\",\"user_mtime\":\"2026-02-27T23:18:05Z\",\"date_type\":\"inclusive\",\"label\":\"creation\",\"jsonmodel_type\":\"date\"}],\"external_documents\":[],\"rights_statements\":[],\"linked_agents\":[],\"import_previous_arks\":[],\"revision_statements\":[],\"instances\":[],\"deaccessions\":[],\"related_accessions\":[],\"classifications\":[],\"notes\":[],\"metadata_rights_declarations\":[],\"uri\":\"/repositories/2/resources/5\",\"repository\":{\"ref\":\"/repositories/2\",\"_resolved\":{\"lock_version\":0,\"repo_code\":\"test_repo_37\",\"name\":\"Test
+        repository 392\",\"created_by\":\"admin\",\"last_modified_by\":\"admin\",\"create_time\":\"2026-02-27T23:18:03Z\",\"system_mtime\":\"2026-02-27T23:18:03Z\",\"user_mtime\":\"2026-02-27T23:18:03Z\",\"publish\":false,\"oai_is_disabled\":false,\"slug\":\"test_repo_37\",\"is_slug_auto\":true,\"position\":1,\"country\":\"US\",\"jsonmodel_type\":\"repository\",\"uri\":\"/repositories/2\",\"display_string\":\"Test
+        repository 392 (test_repo_37)\",\"agent_representation\":{\"ref\":\"/agents/corporate_entities/1\",\"_resolved\":{\"lock_version\":1,\"publish\":false,\"created_by\":\"admin\",\"last_modified_by\":\"admin\",\"create_time\":\"2026-02-27T23:18:03Z\",\"system_mtime\":\"2026-02-27T23:18:08Z\",\"user_mtime\":\"2026-02-27T23:18:03Z\",\"is_slug_auto\":true,\"jsonmodel_type\":\"agent_corporate_entity\",\"agent_contacts\":[{\"lock_version\":0,\"name\":\"test_repo_37\",\"created_by\":\"admin\",\"last_modified_by\":\"admin\",\"create_time\":\"2026-02-27T23:18:03Z\",\"system_mtime\":\"2026-02-27T23:18:03Z\",\"user_mtime\":\"2026-02-27T23:18:03Z\",\"jsonmodel_type\":\"agent_contact\",\"telephones\":[],\"notes\":[],\"is_representative\":false}],\"agent_record_controls\":[],\"agent_alternate_sets\":[],\"agent_conventions_declarations\":[],\"agent_other_agency_codes\":[],\"agent_maintenance_histories\":[],\"agent_record_identifiers\":[],\"agent_identifiers\":[],\"agent_sources\":[],\"agent_places\":[],\"agent_occupations\":[],\"agent_functions\":[],\"agent_topics\":[],\"agent_resources\":[],\"linked_agent_roles\":[],\"external_documents\":[],\"notes\":[],\"used_within_repositories\":[],\"used_within_published_repositories\":[],\"dates_of_existence\":[],\"used_languages\":[],\"metadata_rights_declarations\":[],\"names\":[{\"lock_version\":0,\"primary_name\":\"test_repo_37\",\"sort_name\":\"test_repo_37\",\"sort_name_auto_generate\":true,\"created_by\":\"admin\",\"last_modified_by\":\"admin\",\"create_time\":\"2026-02-27T23:18:03Z\",\"system_mtime\":\"2026-02-27T23:18:03Z\",\"user_mtime\":\"2026-02-27T23:18:03Z\",\"authorized\":true,\"is_display_name\":true,\"jurisdiction\":false,\"conference_meeting\":false,\"source\":\"local\",\"jsonmodel_type\":\"name_corporate_entity\",\"use_dates\":[],\"parallel_names\":[]}],\"related_agents\":[],\"uri\":\"/agents/corporate_entities/1\",\"agent_type\":\"agent_corporate_entity\",\"is_linked_to_published_record\":false,\"display_name\":{\"lock_version\":0,\"primary_name\":\"test_repo_37\",\"sort_name\":\"test_repo_37\",\"sort_name_auto_generate\":true,\"created_by\":\"admin\",\"last_modified_by\":\"admin\",\"create_time\":\"2026-02-27T23:18:03Z\",\"system_mtime\":\"2026-02-27T23:18:03Z\",\"user_mtime\":\"2026-02-27T23:18:03Z\",\"authorized\":true,\"is_display_name\":true,\"jurisdiction\":false,\"conference_meeting\":false,\"source\":\"local\",\"jsonmodel_type\":\"name_corporate_entity\",\"use_dates\":[],\"parallel_names\":[]},\"title\":\"test_repo_37\",\"is_repo_agent\":\"Test
+        repository 392\"}}}},\"caas_next_refid\":{\"lock_version\":20,\"resource_id\":5,\"next_refid\":12,\"create_time\":\"2026-02-27T23:18:05Z\",\"system_mtime\":\"2026-02-27T23:18:07Z\",\"user_mtime\":\"2026-02-27T23:18:07Z\",\"jsonmodel_type\":\"caas_next_refid\",\"uri\":\"/plugins/caas_next_refid/1\"},\"tree\":{\"ref\":\"/repositories/2/resources/5/tree\"}}","years":[1906,1907],"jsonmodel_type":"resource"}],"facets":{"facet_queries":{},"facet_fields":{},"facet_ranges":{},"facet_intervals":{},"facet_heatmaps":{}}}
+
+        '
+    headers:
+      Cache-Control:
+      - private, must-revalidate, max-age=0
+      Content-Length:
+      - '6732'
+      Content-Type:
+      - application/json
+      Date:
+      - Fri, 27 Feb 2026 23:27:53 GMT
+      Server:
+      - Jetty(9.4.44.v20210927)
+      X-Content-Type-Options:
+      - nosniff
+    status:
+      code: 200
+      message: OK
+version: 1
